Fix Ambito-grid text overflow
Every 'card; in abito grid had an issue where if the 'materias' tittles were long the text would overflow outside of the card.
Fixed card height was removed and the order in which js functions are fired (card & masonry) was altered so Auto heights can be achieved and masonry layout would still work.